Add GitHub Actions to auto-update nbdev project pages
This commit introduces a new GitHub Action workflow: gh-pages.yaml. The role of this workflow is to automatically build and push updates to the nbdev project pages upon new pushes or pull requests. This automation is beneficial for keeping the project page up-to-date with the latest changes without manual intervention.