Add introduction markdown to index notebook
A new markdown cell has been added to the top of the index.ipynb notebook. This includes a YAML front matter block with document metadata like title, author, and date, and format configurations. This was done to provide an introductory description and improve the readability and organization of the notebook.